Restructure JS and add HTML
Added HTML structure manually rather than create it with Javascript.
With the Javascript creation, I was unable to remove the event listen
to board grid once a game was won. Now I am able to return the addTic
function and remove it from the board grid so no more moves can be played.
It also looks much more readable.